+++ src/app/profile/[userId]/page.tsx
@@ -1,7 +1,5 @@
 import { Prisma } from '@prisma/client'
-import { revalidatePath } from 'next/cache'
 
-import { CreatedBuildCardActions } from '@/app/profile/[userId]/(components)/CreatedBuildCardActions'
 import { getServerSession } from '@/features/auth/lib'
 import { BuildCard } from '@/features/build/components/build-card/BuildCard'
 import { communityBuildsQuery } from '@/features/build/filters/queries/community-builds'